More work

+++ angular-sanvipop/src/app/products/product-card/product-card.component.ts
@@ -4,8 +4,10 @@ import { Product } from '../interfaces/product';
 import { RouterLink } from '@angular/router';
 import { PostsService } from '../services/posts.service';
 import { FaIconLibrary, FontAwesomeModule } from '@fortawesome/angular-fontawesome';
-import { faEye } from '@fortawesome/free-regular-svg-icons';
-import { faTrash } from '@fortawesome/free-solid-svg-icons';
+import { faEye, faHeart } from '@fortawesome/free-regular-svg-icons';
+import { faTrash, faPenToSquare, faHeart as faHeartSolid } from '@fortawesome/free-solid-svg-icons';
+import { NgbModal } from '@ng-bootstrap/ng-bootstrap';
+import { ConfirmModalComponent } from '../../modals/confirm-modal/confirm-modal.component';
 
 @Component({
   selector: 'product-card',
@@ -20,15 +22,38 @@ export class ProductCardComponent {
 
   #postsService = inject(PostsService);
   #faIconLibrary = inject(FaIconLibrary);
+  #modalService = inject(NgbModal);
 
   constructor() {
-    this.#faIconLibrary.addIcons(faEye, faTrash);
+    this.#faIconLibrary.addIcons(faEye, faTrash, faPenToSquare, faHeart, faHeartSolid);
   }
 
-  deleteProduct() {
-    this.#postsService.deleteProduct(this.product.id).subscribe({
-      next: () => (this.deleted.emit()),
-      error: () => console.error('Error borrando el producto'),
-    });
+  async deleteProduct() {
+    const modalRef = this.#modalService.open(ConfirmModalComponent);
+    modalRef.componentInstance.type = 'warning';
+    modalRef.componentInstance.title = '¿Estás seguro de borrar este producto?';
+    modalRef.componentInstance.body = 'Esta acción no se puede deshacer';
+    
+    if(await modalRef.result.catch(() => false)) {
+      this.#postsService.deleteProduct(this.product.id).subscribe({
+        next: () => this.deleted.emit(),
+        error: () => console.error('Error borrando el producto'),
+      });
+    }
+  }
+
+  changeFavorite() {
+    if(this.product.bookmarked) {
+      this.#postsService.deleteFavorite(this.product.id).subscribe({
+        next: () => console.error('Producto quitado de favorito'),
+        error: () => console.error('Error'),
+      });
+    } else {
+      console.log(this.product.id);
+      this.#postsService.addFavorite(this.product.id).subscribe({
+        next: () => console.error('Producto añadido a favorito'),
+        error: () => console.error('Error'),
+      });
+    }
   }
 }
